feat: separate components & context for advanced
