+++ src/contexts/notification/NotificationProvider.tsx
@@ -1,0 +1,37 @@
+import { PropsWithChildren, useCallback, useMemo, useState } from "react";
+import { INotification, NotificationContext } from "./NotificationContext";
+
+const NotificationProvider: React.FC<PropsWithChildren> = ({ children }) => {
+  const [notifications, setNotifications] = useState<INotification[]>([]);
+
+  const addNotification = useCallback(
+    (message: string, type: INotification["type"]) => {
+      const newNotification: INotification = {
+        id: Date.now(),
+        message,
+        type,
+      };
+      setNotifications((prev) => [...prev, newNotification]);
+    },
+    [],
+  );
+
+  const removeNotification = useCallback((id: number) => {
+    setNotifications((prev) =>
+      prev.filter((notification) => notification.id !== id),
+    );
+  }, []);
+
+  const value = useMemo(
+    () => ({ notifications, addNotification, removeNotification }),
+    [notifications, addNotification, removeNotification],
+  );
+
+  return (
+    <NotificationContext.Provider value={value}>
+      {children}
+    </NotificationContext.Provider>
+  );
+};
+
+export default NotificationProvider;
