+++ src/contexts/notification/useNotificationContext.tsx
@@ -1,0 +1,16 @@
+import { useContext } from "react";
+import { NotificationContext } from "./NotificationContext";
+
+const useNotificationContext = () => {
+  const context = useContext(NotificationContext);
+
+  if (!context) {
+    throw new Error(
+      "useNotification must be used within a NotificationProvider",
+    );
+  }
+
+  return context;
+};
+
+export default useNotificationContext;
