+++ src/contexts/theme/ThemeContext.tsx
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+export type Theme = "light" | "dark";
+
+interface ThemeContextType {
+  theme: Theme;
+  toggleTheme: () => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType | null>(null);
