+++ src/contexts/theme/ThemeProvider.tsx
@@ -1,0 +1,24 @@
+import { PropsWithChildren, useCallback, useMemo, useState } from "react";
+import { Theme, ThemeContext } from "./ThemeContext";
+
+const ThemeProvider: React.FC<PropsWithChildren> = ({ children }) => {
+  const [theme, setTheme] = useState<Theme>("light");
+
+  const toggleTheme = useCallback(() => {
+    setTheme((prev) => (prev === "light" ? "dark" : "light"));
+  }, []);
+
+  const value = useMemo(() => ({ theme, toggleTheme }), [theme, toggleTheme]);
+
+  return (
+    <ThemeContext.Provider value={value}>
+      <div
+        className={`min-h-screen ${theme === "light" ? "bg-gray-100" : "bg-gray-900 text-white"}`}
+      >
+        {children}
+      </div>
+    </ThemeContext.Provider>
+  );
+};
+
+export default ThemeProvider;
