+++ src/contexts/user/UserProvider.tsx
@@ -1,0 +1,22 @@
+import { PropsWithChildren, useCallback, useMemo, useState } from "react";
+import { IUser, UserContext } from "./UserContext";
+
+const UserProvider: React.FC<PropsWithChildren> = ({ children }) => {
+  const [user, setUser] = useState<IUser | null>(null);
+
+  const login = useCallback((email: string) => {
+    setUser({ id: 1, name: "홍길동", email });
+    // addNotification("성공적으로 로그인되었습니다", "success");
+  }, []);
+
+  const logout = useCallback(() => {
+    setUser(null);
+    // addNotification("로그아웃되었습니다", "info");
+  }, []);
+
+  const value = useMemo(() => ({ user, login, logout }), [user, login, logout]);
+
+  return <UserContext.Provider value={value}>{children}</UserContext.Provider>;
+};
+
+export default UserProvider;
