+++ src/contexts/user/useUserContext.tsx
@@ -1,0 +1,14 @@
+import { useContext } from "react";
+import { UserContext } from "./UserContext";
+
+const useUserContext = () => {
+  const context = useContext(UserContext);
+
+  if (!context) {
+    throw new Error("useUser must be used within a UserProvider");
+  }
+
+  return context;
+};
+
+export default useUserContext;
